--- futures/bin/dailyNewsDigest.txt
@@ -1,1 +1,0 @@
-<Contents of dailyNewsDigest.txt>
