--- futures/bin/futures1.dart
@@ -1,25 +1,0 @@
-import 'dart:io';
-
-void printDailyNewsDigest() {
-  File file = new File("dailyNewsDigest.txt");
-  print(file.readAsStringSync());
-}
-
-void main() {
-  printDailyNewsDigest();
-  printWinningLotteryNumbers();
-  printWeatherForecast();
-  printBaseballScore();
-}
-
-printWinningLotteryNumbers() {
-  print('Winning lotto numbers: [23, 63, 87, 26, 2]');
-}
-
-printWeatherForecast() {
-  print('Tomorrow\'s forecast: 70F, sunny.');
-}
-
-printBaseballScore() {
-  print('Baseball score: Red Sox 10, Yankees 0');
-}
