+++ futures/futures-api-catch-error/pubspec.yaml
@@ -1,0 +1,6 @@
+name: read_file_futures_catch_errors
+description: A sample application that reads a file on dartlang using the Futures API, and prints the contents; errors are caught.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+gist: https://gist.github.com/Sfshaza/fc9930b9ccc4427808d0
+dartpad: https://dartpad.dartlang.org/fc9930b9ccc4427808d0
