+++ futures/futures-api/pubspec.yaml
@@ -1,0 +1,6 @@
+name: read_file_futures
+description: A sample application that reads a file on dartlang using the Futures API, and prints the contents.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+gist: https://gist.github.com/Sfshaza/c81150d1e64b48b596a9
+dartpad: https://dartpad.dartlang.org/c81150d1e64b48b596a9
