+++ futures/sequential-processing/pubspec.yaml
@@ -1,0 +1,6 @@
+name: sequential_procesing
+description: A sample application that performs time consuming operations sequentially.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+gist: https://gist.github.com/Sfshaza/d3d5fb115c8d33e61c94
+dartpad: https://dartpad.dartlang.org/d3d5fb115c8d33e61c94
