+++ futures/sequential-processing/web/main.dart
@@ -1,0 +1,20 @@
+// Copyright (c) 2013, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:async';
+
+void main() {
+  expensiveA()
+      .then((aValue) => expensiveB())
+      .then((bValue) => expensiveC())
+      .then((cValue) => doSomethingWith(cValue));
+}
+
+Future expensiveA() => new Future.value('from expensiveA');
+Future expensiveB() => new Future.value('from expensiveB');
+Future expensiveC() => new Future.value('from expensiveC');
+
+doSomethingWith(value) {
+  print(value);
+}
