+++ futures/waiting-on-futures/pubspec.yaml
@@ -1,0 +1,6 @@
+name: waiting_on_futures
+description: A sample application that triggers three functions and waits until all are finished.
+
+homepage: https://www.dartlang.org/docs/tutorials/futures/
+gist: https://gist.github.com/Sfshaza/4dd5d8ee6a56e1099f5c
+dartpad: https://dartpad.dartlang.org/4dd5d8ee6a56e1099f5c
